# v6: restore missed logging messages