--- a/v6/logger.go
+++ b/v6/logger.go
@@ -6,16 +6,16 @@ import (
 
 // Define the logrus log levels
 const (
-	LogLevelPanic = LogLevel(logrus.PanicLevel)
-	LogLevelFatal = LogLevel(logrus.FatalLevel)
-	LogLevelError = LogLevel(logrus.ErrorLevel)
-	LogLevelWarn  = LogLevel(logrus.WarnLevel)
-	LogLevelInfo  = LogLevel(logrus.InfoLevel)
-	LogLevelDebug = LogLevel(logrus.DebugLevel)
-	LogLevelTrace = LogLevel(logrus.TraceLevel)
+	LogLevelPanic = logrus.PanicLevel
+	LogLevelFatal = logrus.FatalLevel
+	LogLevelError = logrus.ErrorLevel
+	LogLevelWarn  = logrus.WarnLevel
+	LogLevelInfo  = logrus.InfoLevel
+	LogLevelDebug = logrus.DebugLevel
+	LogLevelTrace = logrus.TraceLevel
 )
 
-type LogLevel uint32
+type LogLevel = logrus.Level
 
 // Logger defines the interface this library logs with.
 type Logger interface {
@@ -35,9 +35,29 @@ type Logger interface {
 	Errorln(args ...interface{})
 }
 
+// LoggerWithLevel is optionally implemented by a Logger.
+// It is notably implemented by logrus.Logger and thus
+// by the DefaultLogger returned by this package.
+type LoggerWithLevel interface {
+	// GetLevel returns the current logging level.
+	GetLevel() LogLevel
+}
+
 // DefaultLogger creates the default logger with specified log level (logrus.New()).
 func DefaultLogger(level LogLevel) Logger {
 	logger := logrus.New()
-	logger.SetLevel(logrus.Level(level))
+	logger.SetLevel(level)
 	return logger
+}
+
+// leveledLogger wraps a Logger for efficiency reasons: it's a static type
+// rather than an interface so the compiler can inline the level check
+// and thus avoid the allocation for the arguments.
+type leveledLogger struct {
+	level LogLevel
+	Logger
+}
+
+func (log *leveledLogger) enabled(level LogLevel) bool {
+	return level <= log.level
 }