--- a/v6/rollout_integration_test.go
+++ b/v6/rollout_integration_test.go
@@ -27,16 +27,21 @@ type integrationTest struct {
 func BenchmarkGetValue(b *testing.B) {
 	b.ReportAllocs()
 	logger := DefaultLogger(LogLevelError)
-	client := NewCustomClient(integrationTests[0].sdkKey, ClientConfig{Logger: logger, Mode: LazyLoad(120, true)})
+	client := NewCustomClient(integrationTests[0].sdkKey, ClientConfig{
+		Logger:         logger,
+		Mode:           ManualPoll(),
+		StaticLogLevel: true,
+	})
 	client.Refresh()
 	defer client.Close()
-	b.ResetTimer()
-	val := client.GetValueForUser("bool30TrueAdvancedRules", nil, nil)
-	if val != true {
+	user := NewUser("unknown-identifier")
+	val := client.GetValueForUser("bool30TrueAdvancedRules", "default", user)
+	if val != false {
 		b.Fatalf("unexpected result %#v", val)
 	}
+	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		client.GetValueForUser("bool30TrueAdvancedRules", nil, nil)
+		client.GetValueForUser("bool30TrueAdvancedRules", "default", user)
 	}
 }
 
@@ -86,6 +91,7 @@ func (test integrationTest) runTest(t *testing.T) {
 		cfg = srv.config()
 	}
 	cfg.Mode = test.mode
+	cfg.Logger = newTestLogger(t, LogLevelError)
 	client := NewCustomClient(test.sdkKey, cfg)
 	client.Refresh()
 	defer client.Close()