v6: restore missed logging messages

There were a few informational and warning messages that had
been missed in the transition. We reinstate those, but they
made things slower again, so we add  wrapper type
that can be used to inspect the logging level before logging,
which avoids the allocations from calling the log methods.

Also change the log level type to be an alias to the logrus
level type so that we can define the `LoggerWithLevel` interface
without needing to use two level types in the public API.
This should be backwardly compatible except for some
highly unlikely (and easily fixable) use cases, such as `fmt.Sprint(level)`
formatting differently because `logrus.Level` implements a `String` method.

We also use a more realistic benchmark that actually
exercises the rules by passing a User value in.

Change from before the evaluator preprocessing to this PR:

```
name        old time/op    new time/op    delta
GetValue-8     590ns ± 9%     283ns ± 5%  -52.10%  (p=0.008 n=5+5)

name        old alloc/op   new alloc/op   delta
GetValue-8      320B ± 0%       96B ± 0%  -70.00%  (p=0.008 n=5+5)

name        old allocs/op  new allocs/op  delta
GetValue-8      10.0 ± 0%       2.0 ± 0%  -80.00%  (p=0.008 n=5+5)
```

Change from the most recent master to this PR:

```
name        old time/op    new time/op    delta
GetValue-8     307ns ± 3%     283ns ± 5%   -8.04%  (p=0.008 n=5+5)

name        old alloc/op   new alloc/op   delta
GetValue-8      112B ± 0%       96B ± 0%  -14.29%  (p=0.008 n=5+5)

name        old allocs/op  new allocs/op  delta
GetValue-8      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.008 n=5+5)
```